chore: renamed chats to chat page
