+++ src/components/Chat/Landing.tsx
@@ -71,7 +71,7 @@ function Landing() {
           }}
         >
           <Image
-            src={require("@/pages/chats/images/image_1.png")}
+            src={require("@/pages/chat/images/image_1.png")}
             alt={"left"}
             priority={true}
             fill
@@ -92,7 +92,7 @@ function Landing() {
           }}
         >
           <Image
-            src={require("@/pages/chats/images/image_2.png")}
+            src={require("@/pages/chat/images/image_2.png")}
             alt={"middle"}
             priority={true}
             fill
@@ -119,7 +119,7 @@ function Landing() {
           }}
         >
           <Image
-            src={require("@/pages/chats/images/image_3.png")}
+            src={require("@/pages/chat/images/image_3.png")}
             alt={"right"}
             priority={true}
             fill
